requirements.txt
requests
anyio==4.12.1
APScheduler==3.11.2
certifi==2026.2.25
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==26.0
psycopg[binary]==3.2.10
python-dotenv==1.2.2
python-telegram-bot==22.6
setuptools==82.0.0
sniffio==1.3.1
tzdata==2025.3
tzlocal==5.3.1
wheel==0.46.3
